+++ kernels/vlasov/vlasov_mom_kernels_ser_1x3v.c
@@ -1,19 +1,19 @@
 #include <gkyl_vlasov_mom_kernels.h>
 
-void vlasov_mom_1x3v_m0_p1(const double *w, const double *dxv, const long *idx, const double *f, double* restrict out) 
+void vlasov_mom_1x3v_m0_p1(const double *w, const double *dxv, const int *idx, const double *f, double* restrict out) 
 { 
   const double volFact = dxv[1]*dxv[2]*dxv[3]/8; 
   out[0] += 2.828427124746191*f[0]*volFact; 
   out[1] += 2.828427124746191*f[1]*volFact; 
 } 
-void vlasov_mom_1x3v_m0_p2(const double *w, const double *dxv, const long *idx, const double *f, double* restrict out) 
+void vlasov_mom_1x3v_m0_p2(const double *w, const double *dxv, const int *idx, const double *f, double* restrict out) 
 { 
   const double volFact = dxv[1]*dxv[2]*dxv[3]/8; 
   out[0] += 2.828427124746191*f[0]*volFact; 
   out[1] += 2.828427124746191*f[1]*volFact; 
   out[2] += 2.828427124746191*f[11]*volFact; 
 } 
-void vlasov_mom_1x3v_m1i_p1(const double *w, const double *dxv, const long *idx, const double *f, double* restrict out) 
+void vlasov_mom_1x3v_m1i_p1(const double *w, const double *dxv, const int *idx, const double *f, double* restrict out) 
 { 
   const double volFact = dxv[1]*dxv[2]*dxv[3]/8; 
   const double wx1 = w[1], dv1 = dxv[1]; 
@@ -26,7 +26,7 @@ void vlasov_mom_1x3v_m1i_p1(const double *w, const double *dxv, const long *idx,
   out[4] += volFact*(2.828427124746191*f[0]*wx3+0.8164965809277261*f[4]*dv3); 
   out[5] += volFact*(2.828427124746191*f[1]*wx3+0.8164965809277261*f[8]*dv3); 
 } 
-void vlasov_mom_1x3v_m1i_p2(const double *w, const double *dxv, const long *idx, const double *f, double* restrict out) 
+void vlasov_mom_1x3v_m1i_p2(const double *w, const double *dxv, const int *idx, const double *f, double* restrict out) 
 { 
   const double volFact = dxv[1]*dxv[2]*dxv[3]/8; 
   const double wx1 = w[1], dv1 = dxv[1]; 
@@ -42,7 +42,7 @@ void vlasov_mom_1x3v_m1i_p2(const double *w, const double *dxv, const long *idx,
   out[7] += volFact*(2.828427124746191*f[1]*wx3+0.8164965809277261*f[8]*dv3); 
   out[8] += volFact*(2.828427124746191*f[11]*wx3+0.816496580927726*f[25]*dv3); 
 } 
-void vlasov_mom_1x3v_m2ij_p1(const double *w, const double *dxv, const long *idx, const double *f, double* restrict out) 
+void vlasov_mom_1x3v_m2ij_p1(const double *w, const double *dxv, const int *idx, const double *f, double* restrict out) 
 { 
   const double volFact = dxv[1]*dxv[2]*dxv[3]/8; 
   const double wx1 = w[1], dv1 = dxv[1]; 
@@ -64,7 +64,7 @@ void vlasov_mom_1x3v_m2ij_p1(const double *w, const double *dxv, const long *idx
   out[10] += volFact*(2.828427124746191*f[0]*wx3_sq+1.632993161855453*f[4]*dv3*wx3+0.2357022603955158*f[0]*dv3_sq); 
   out[11] += volFact*(2.828427124746191*f[1]*wx3_sq+1.632993161855453*f[8]*dv3*wx3+0.2357022603955158*f[1]*dv3_sq); 
 } 
-void vlasov_mom_1x3v_m2ij_p2(const double *w, const double *dxv, const long *idx, const double *f, double* restrict out) 
+void vlasov_mom_1x3v_m2ij_p2(const double *w, const double *dxv, const int *idx, const double *f, double* restrict out) 
 { 
   const double volFact = dxv[1]*dxv[2]*dxv[3]/8; 
   const double wx1 = w[1], dv1 = dxv[1]; 
@@ -92,7 +92,7 @@ void vlasov_mom_1x3v_m2ij_p2(const double *w, const double *dxv, const long *idx
   out[16] += volFact*(2.828427124746191*f[1]*wx3_sq+1.632993161855453*f[8]*dv3*wx3+0.2108185106778921*f[28]*dv3_sq+0.2357022603955158*f[1]*dv3_sq); 
   out[17] += volFact*(2.828427124746191*f[11]*wx3_sq+1.632993161855453*f[25]*dv3*wx3+0.2357022603955158*f[11]*dv3_sq); 
 } 
-void vlasov_mom_1x3v_m2_p1(const double *w, const double *dxv, const long *idx, const double *f, double* restrict out) 
+void vlasov_mom_1x3v_m2_p1(const double *w, const double *dxv, const int *idx, const double *f, double* restrict out) 
 { 
   const double volFact = dxv[1]*dxv[2]*dxv[3]/8; 
   const double wx1 = w[1], dv1 = dxv[1]; 
@@ -104,7 +104,7 @@ void vlasov_mom_1x3v_m2_p1(const double *w, const double *dxv, const long *idx, 
   out[0] += volFact*(2.828427124746191*f[0]*wx3_sq+1.632993161855453*f[4]*dv3*wx3+2.828427124746191*f[0]*wx2_sq+1.632993161855453*f[3]*dv2*wx2+2.828427124746191*f[0]*wx1_sq+1.632993161855453*f[2]*dv1*wx1+0.2357022603955158*f[0]*dv3_sq+0.2357022603955158*f[0]*dv2_sq+0.2357022603955158*f[0]*dv1_sq); 
   out[1] += volFact*(2.828427124746191*f[1]*wx3_sq+1.632993161855453*f[8]*dv3*wx3+2.828427124746191*f[1]*wx2_sq+1.632993161855453*f[6]*dv2*wx2+2.828427124746191*f[1]*wx1_sq+1.632993161855453*f[5]*dv1*wx1+0.2357022603955158*f[1]*dv3_sq+0.2357022603955158*f[1]*dv2_sq+0.2357022603955158*f[1]*dv1_sq); 
 } 
-void vlasov_mom_1x3v_m2_p2(const double *w, const double *dxv, const long *idx, const double *f, double* restrict out) 
+void vlasov_mom_1x3v_m2_p2(const double *w, const double *dxv, const int *idx, const double *f, double* restrict out) 
 { 
   const double volFact = dxv[1]*dxv[2]*dxv[3]/8; 
   const double wx1 = w[1], dv1 = dxv[1]; 
@@ -117,7 +117,7 @@ void vlasov_mom_1x3v_m2_p2(const double *w, const double *dxv, const long *idx, 
   out[1] += volFact*(2.828427124746191*f[1]*wx3_sq+1.632993161855453*f[8]*dv3*wx3+2.828427124746191*f[1]*wx2_sq+1.632993161855453*f[6]*dv2*wx2+2.828427124746191*f[1]*wx1_sq+1.632993161855453*f[5]*dv1*wx1+0.2108185106778921*f[28]*dv3_sq+0.2357022603955158*f[1]*dv3_sq+0.2108185106778921*f[23]*dv2_sq+0.2357022603955158*f[1]*dv2_sq+0.2108185106778921*f[20]*dv1_sq+0.2357022603955158*f[1]*dv1_sq); 
   out[2] += volFact*(2.828427124746191*f[11]*wx3_sq+1.632993161855453*f[25]*dv3*wx3+2.828427124746191*f[11]*wx2_sq+1.632993161855453*f[21]*dv2*wx2+2.828427124746191*f[11]*wx1_sq+1.632993161855453*f[19]*dv1*wx1+0.2357022603955158*f[11]*dv3_sq+0.2357022603955158*f[11]*dv2_sq+0.2357022603955158*f[11]*dv1_sq); 
 } 
-void vlasov_mom_1x3v_m3i_p1(const double *w, const double *dxv, const long *idx, const double *f, double* restrict out) 
+void vlasov_mom_1x3v_m3i_p1(const double *w, const double *dxv, const int *idx, const double *f, double* restrict out) 
 { 
   const double volFact = dxv[1]*dxv[2]*dxv[3]/8; 
   const double wx1 = w[1], dv1 = dxv[1]; 
@@ -136,7 +136,7 @@ void vlasov_mom_1x3v_m3i_p1(const double *w, const double *dxv, const long *idx,
   out[4] += volFact*(2.828427124746191*f[0]*wx3*wx3_sq+2.449489742783178*f[4]*dv3*wx3_sq+2.828427124746191*f[0]*wx2_sq*wx3+1.632993161855453*f[3]*dv2*wx2*wx3+2.828427124746191*f[0]*wx1_sq*wx3+1.632993161855453*f[2]*dv1*wx1*wx3+0.7071067811865475*f[0]*dv3_sq*wx3+0.2357022603955158*f[0]*dv2_sq*wx3+0.2357022603955158*f[0]*dv1_sq*wx3+0.8164965809277261*f[4]*dv3*wx2_sq+0.4714045207910317*f[10]*dv2*dv3*wx2+0.8164965809277261*f[4]*dv3*wx1_sq+0.4714045207910317*f[9]*dv1*dv3*wx1+0.1224744871391589*f[4]*dv3*dv3_sq+0.06804138174397717*f[4]*dv2_sq*dv3+0.06804138174397717*f[4]*dv1_sq*dv3); 
   out[5] += volFact*(2.828427124746191*f[1]*wx3*wx3_sq+2.449489742783178*f[8]*dv3*wx3_sq+2.828427124746191*f[1]*wx2_sq*wx3+1.632993161855453*f[6]*dv2*wx2*wx3+2.828427124746191*f[1]*wx1_sq*wx3+1.632993161855453*f[5]*dv1*wx1*wx3+0.7071067811865475*f[1]*dv3_sq*wx3+0.2357022603955158*f[1]*dv2_sq*wx3+0.2357022603955158*f[1]*dv1_sq*wx3+0.8164965809277261*f[8]*dv3*wx2_sq+0.4714045207910317*f[13]*dv2*dv3*wx2+0.8164965809277261*f[8]*dv3*wx1_sq+0.4714045207910317*f[12]*dv1*dv3*wx1+0.1224744871391589*f[8]*dv3*dv3_sq+0.06804138174397717*f[8]*dv2_sq*dv3+0.06804138174397717*f[8]*dv1_sq*dv3); 
 } 
-void vlasov_mom_1x3v_m3i_p2(const double *w, const double *dxv, const long *idx, const double *f, double* restrict out) 
+void vlasov_mom_1x3v_m3i_p2(const double *w, const double *dxv, const int *idx, const double *f, double* restrict out) 
 { 
   const double volFact = dxv[1]*dxv[2]*dxv[3]/8; 
   const double wx1 = w[1], dv1 = dxv[1];
